requirements.txt
pyspark==3.5.1
streamlit
uvicorn[standard]
fastapi
psycopg2-binary
pandas
requests
scikit-learn
apache-airflow-providers-fab
yfinance
google-generativeai
